==> Project_2/02_imdb_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
import json

### Cleaning the IMDB Data

Initially, I had planned to iterate through and scrape the individual pages for film and director.  To this end, I selected the unique IMDB ids for director and title from the urls present on the title search page.  I also separated out the three genre keywords, intending to use all three columns as part of the analysis. Ultimately, I used only the first of the three keywords, and I used the aggregate information for each director within the dataset for the film counts for each.

Rather than cleaning and formatting each string or data type as it came through the for loops in the data scraping, I cleaned and formatted them here using the Pandas series.

imdb_df2 = pd.read_json('imdb_data2.json', lines=True)
# lines=True : read file as a json object per line

imdb_df2.head()

#### Segmenting Out Unique IDs for Title and Director

The second segment of the title url (`title_link`) is the unique `imdb_id` for the film, facilitating specific title searching on both IMDB and Box Office Mojo.  The unique link url for each director had a similar ID.

imdb_df2['imdb_id'] = (imdb_df2['title_link'].str.split("/", n=3, expand=True))[2]                               
imdb_df2['imdb_id']

imdb_df2['director_id'] = (imdb_df2['director_link'].str.split("/", n=3, expand=True))[2]                               
imdb_df2['director_id']

imdb_df2.head()

#### Splitting Up Genre Keywords Into Individual Features

imdb_df2[['genre_01', 'genre_02', 'genre_03']] = imdb_df2['imdb_genre_list'].str.split(", ", n=2, expand=True)

imdb_df2[['genre_01', 'genre_02', 'genre_03']].head()

imdb_df2['genre_01'] = imdb_df2['genre_01'].str.strip("\n")

#### String Cleaning, Type Standardization, Dropping of Null Values

imdb_df2['runtime'] = imdb_df2['runtime'].str.strip(' min')
imdb_df2['runtime']= imdb_df2['runtime'].replace('', np.NaN)
imdb_df2['runtime'] = pd.to_numeric(imdb_df2['runtime'], errors='coerce')

imdb_df2.head()

####  Formatting Column Labels

imdb_df.columns

imdb_df2 = imdb_df2[['title', 'imdb_id', 'title_link', 'year', 'runtime', 'mpaa_rating', 'director_name', 'director_id', 'director_link', 'genre_01', 'genre_02', 'genre_03','imdb_rating', 'votes', 'metascore']]

imdb_df2['runtime'].unique()

imdb_df2['votes']= imdb_df2['votes'].replace('', np.NaN)
imdb_df2['votes'] = pd.to_numeric(imdb_df2['votes'], errors='coerce')

imdb_df2['votes'].unique()

imdb_df2.head()

imdb_df2['year'] = imdb_df2['year'].str.strip('[)XIV( ]')

imdb_df2['year'].unique()

imdb_df2['year']= imdb_df2['year'].replace('', np.NaN)

imdb_df2['year'].unique()

imdb_df2.info()

imdb_df2.sample(10, random_state=42)

imdb_df2.describe()

imdb_df2['year'].value_counts()

#### Dropping Less Common MPAA Ratings Values

imdb_df2['mpaa_rating'].value_counts()
# let's drop some of these ratings, esp TV and X

imdb_df2['mpaa_rating'].unique()

imdb_df2['genre_01'].value_counts()

imdb_df2['genre_02'].value_counts()

imdb_df2['genre_03'].value_counts()

imdb_df2['genre_02']= imdb_df2['genre_02'].str.strip()

imdb_df2['genre_02'].value_counts()

imdb_df2['genre_03']= imdb_df2['genre_03'].str.strip()

imdb_df2['genre_03'].value_counts()

imdb_df2['genre_01']= imdb_df2['genre_01'].str.strip()

imdb_df2['genre_01'].value_counts()

imdb_df2.columns

imdb_lookup_df3 = imdb_df2_ms[['title', 'imdb_id','title_link', 'director_name', 'director_id', 'director_link']]

####  Creating A Title String for URL Searching Rotten Tomatoes

imdb_lookup_df3['title_cc'] = imdb_df2_ms['title'].str.lower() 

imdb_lookup_df3['title_cc'].head()

imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace('&','and',regex=False)
imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace('\\','_',regex=True)
imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace('episode\s([ivx]*)\s-\s','',regex=True)
imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace("\s",'_',regex=True)
imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace("\W",'',regex=True)
imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.replace('_+','_',regex=True)

print(imdb_lookup_df3.shape)
imdb_lookup_df3.head()

imdb_lookup_df2.shape

imdb_df2['director_name']= imdb_df2['director_name'].replace('NaN', np.NaN)

imdb_df2['director_id']= imdb_df2['director_id'].replace('NaN', np.NaN)

imdb_df2['director_id']= imdb_df2['director_id'].replace('None', np.NaN)

imdb_df2[imdb_df2.director_id.isna()]

imdb_df2['director_link']= imdb_df2['director_link'].replace('NaN', np.NaN)

imdb_df2['director_link']= imdb_df2['director_link'].replace('None', np.NaN)

imdb_df[imdb_df.director_link.isna()]

imdb_df2 = imdb_df2.replace('NaN', np.NaN)

imdb_df2 = imdb_df2.replace('None', np.NaN)

imdb_df2.isna().sum()

imdb_df2_ms = imdb_df2[imdb_df2['metascore'].notna()]

imdb_df2_ms.shape

imdb_df2_ms.describe()

imdb_df2_ms.info()

imdb_df2_ms.head()

imdb_df2_ms.to_json('imdb_data_cleaned_2.json', orient='records', lines=True)

imdb_lookup_df3.to_json('imdb_lookup_df3.json', orient='records', lines=True)
