==> Project_2/03_web_scraping_tomatoes.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import os, re, sys, pickle, requests, resource, json
import time
from random import random
from IPython.core.display import clear_output

from selenium import webdriver
from selenium.webdriver.common.keys import Keys

chromedriver = "/Applications/chromedriver"  # path to the chromedriver executable
os.environ['webdriver.chrome.driver'] = chromedriver
driver = webdriver.Chrome(chromedriver)

## Webscraping Rotten Tomatoes

In order to acquire Tomatometer data, it was necessary to search out the url for each film title individually.  To do this, I imported the cleaned IMDB data and formatted a title string to append to the base Rotten Tomatoes url.  

imdb_lookup_df3 = pd.read_json("imdb_lookup_df3.json", lines=True)

imdb_lookup_df3.shape

imdb_lookup_df3.columns

imdb_lookup_df3['title_cc'] = imdb_lookup_df3['title_cc'].str.strip(':')

title_list = imdb_lookup_df3['title_cc']

title_list[0:20]

# rotten tomatoes url - takes title, lower case, underscores
tomato_url = "https://www.rottentomatoes.com/m/{}"

movie_html = []
tomato_dict = {}
tomato_list = []


def get_page_source(title_list):
    t_success = 0
    t_nan = 0
    
    headers = ['title', 'tomatometer', 'tomato_ct', 'aud_score', 'aud_score_ct']
    
    start_time = time.time()
    tries = 0
    
    for title in title_list:
        url = tomato_url.format(title)
        user_agent = {'User-agent': 'Mozilla/5.0'}
        driver = webdriver.Chrome(chromedriver)
        tomato_response = driver.get(url)
        movie_html.append(driver.page_source)
        
        tries += 1
        time.sleep(2+2*random())
        elapsed_time = time.time() - start_time
        print('Request: {}; Frequency: {} requests/s'.format(tries, tries/elapsed_time))
        clear_output(wait = True)

        tomato_soup = BeautifulSoup(driver.page_source, 'html5lib')
       
        percents = tomato_soup.find_all('span', class_ = 'mop-ratings-wrap__percentage')
        if not percents:
            pass
        try:
            tomatometer = percents[0].text.strip().strip('%')
        except:
            tomatometer = np.nan
        try:
            aud_score = percents[1].text.strip().strip('%')
        except:
            aud_score = np.nan
                
        aud_score_cts = tomato_soup.find_all('strong', class_='mop-ratings-wrap__text--small')
        if not aud_score_cts:
            pass

        try:
            tomato_ct = tomato_soup.find('small', class_="mop-ratings-wrap__text--small").text.strip()
        except:
            tomato_ct = np.nan
        try:
            aud_score_ct = aud_score_cts[1].text.strip('User Ratings: ').strip()
        except:
            aud_score_ct = np.nan

        tomato_dict = dict(zip(headers, [title, tomatometer,tomato_ct,aud_score, aud_score_ct]))

        tomato_list.append(tomato_dict)



get_page_source(title_list)

tomato_list[0:10]

len(tomato_list)

tomato_df2 = pd.DataFrame(tomato_list)
tomato_df2.head()

len(tomato_df2)

tomato_df2['tomato_ct']= tomato_df2['tomato_ct'].replace('N/A', np.NaN)

tomato_df2['aud_score_ct']= tomato_df2['aud_score_ct'].replace('Not yet availabl', np.NaN)

tomato_df2.describe()

tomato_df2.to_json('tomato_data2.json', orient='records', lines=True)
