==> Project_2/04_web_scraping_bomojo.ipynb <==
from bs4 import BeautifulSoup
import requests
import numpy as np
import pandas as pd

import os, re, sys, pickle, requests, resource, json
import time
from random import random
from IPython.core.display import clear_output

### Webscraping Box Office Mojo

Since I encountered some difficulties in obtaining director information for all of the film entries on IMDB, in particular for films which had more than one director, I attempted to scrape this information from [Box Office Mojo](https://www.boxofficemojo.com/), which had a more approachable tabular format. I used the `imdb_id` field sliced from the `title_link`  from the cleaned IMDB dataframe to do this.

I also attempted to scrape other cast and crew data, but ultimately abandoned this portion due to time constraints.

imdb_lookup_df3 = pd.read_json("imdb_lookup_df3.json", lines=True)

imdb_lookup_df3.head()

id_list = imdb_lookup_df3['imdb_id']

def get_movie_value(soup, field_name):
    obj = soup.find(text = re.compile(field_name))
    
    if not obj:
        return None
    
    next_element = obj.findNext()
    
    if next_element:
        return next_element.text
    else:
        
        return None

mojo_data_list = []

def get_movie_dict(id_list):

    mojo_url = "https://www.boxofficemojo.com/title/{}/?ref_=bo_se_r_1"

# preparing to monitor
    start_time = time.time()
    tries =0

    for movie_id in id_list:
    
        response = requests.get(mojo_url.format(movie_id))
        page = response.text
        
        time.sleep(0.5+2*random())
        tries +=1
        elapsed_time = time.time() - start_time
        print('Request:{}; Frequency: {} requests/s'.format(tries, tries/elapsed_time))
        clear_output(wait = True)
        
        if tries > 2200:
            return "Number of requests exceeded"
            break

        mojo_soup = BeautifulSoup(page, "lxml")

        headers = ["id","title_string","dom_gross", "release_date"]

        title_string = mojo_soup.find('title').text
        
        try:
            domestic_gross = (mojo_soup.find(
                class_='mojo-performance-summary-table')\
                              .find_all('span', class_='money')[0].text)
        except:
            domestic_gross = np.nan
                         
        release_date = get_movie_value(mojo_soup, 'Release Date')
        
        mojo_dict = dict(zip(headers, [movie_id, title_string,
                                       domestic_gross,
                                       release_date]))
        mojo_data_list.append(mojo_dict)

crew_id_dict ={}
cast_id_dict ={}

def get_movie_cast_n_crew(id_list):

    
    mojo_crew_url = "https://www.boxofficemojo.com/title/{}/credits/?ref_=bo_tt_tab#tabs"

# preparing to monitor
    start_time = time.time()
    tries =0

    for movie_id in id_list:
    
        response = requests.get(mojo_crew_url.format(movie_id))
        page = response.text
        
        time.sleep(0.5+2*random())
        tries +=1
        elapsed_time = time.time() - start_time
        print('Request:{}; Frequency: {} requests/s'.format(tries, tries/elapsed_time))
        clear_output(wait = True)
        
        if tries > 2200:
            return "Number of requests exceeded"
            break

        mojo_crew_soup = BeautifulSoup(page, "lxml")
        
        crew_table = mojo_crew_soup.find('table', attrs={'id':'principalCrew'})
        
        crew_rows = [row for row in crew_table.find_all('tr')]
        
     
        crew_list = [movie_id]
        crew_header = ['movie_id']
                
        for row in crew_rows[1:3]:
            crew = row.find_all('td')
            crew_name = crew[0].find('a').text
            crew_list.append(crew_name)
            crew_role = crew[-1].text
            crew_header.append(crew_role)
        
        while crew_list < 4:
            crew_list.append(np.nan)
            
        crew_dict = dict(zip(crew_header, crew_list))

            
            
    mojo_cast_url = "https://www.boxofficemojo.com/title/{}/credits/?ref_=bo_tt_tab#tabs"

    for id in id_list:
        response = requests.get(mojo_cast_url.format(id))
        page = response.text

        mojo_cast_soup = BeautifulSoup(page, "lxml")

        cast_table = mojo_cast_soup.find('table', attrs={'id':'principalCast'})
        
        cast_rows = [row for row in cast_table.find_all('tr')]
        
        crew_list = [movie_id]
        crew_header = ['movie_id', 'actor_01', 'actor_02', 'actor_03']
        
        for row in crew_rows[1:3]:
            crew = row.find_all('td')
            crew_name = crew[0].find('a').text
            crew_list.append(crew_name)
        
        while crew_list < 4:
            crew_list.append(np.nan)
            
               imdb_dict = dict(zip(headers, [title, title_link, mpaa_rating, year, runtime, imdb_rating, metascore,
                              imdb_votes, director_name, director_link, imdb_genres]))

            
        cast_id_dict[id] = cast_dict 

get_movie_dict(id_list)

len(mojo_data_list)

mojo_crew_url = "https://www.boxofficemojo.com/title/{}/credits/?ref_=bo_tt_tab#tabs"
    
# preparing to monitor
start_time = time.time()
tries =0

for movie_id in id_list:

    response = requests.get(mojo_crew_url.format(movie_id))
    page = response.text

    time.sleep(0.5+2*random())
    tries +=1
    elapsed_time = time.time() - start_time
    print('Request:{}; Frequency: {} requests/s'.format(tries, tries/elapsed_time))
    clear_output(wait = True)

    if tries > 2200:
        print("Number of requests exceeded")
        break

    mojo_crew_soup = BeautifulSoup(page, "lxml")

    crew_table = mojo_crew_soup.find('table', attrs={'id':'principalCrew'})

    crew_rows = [row for row in crew_table.find_all('tr')]
    print(crew_rows)

    crew_dict = {}

    for row in crew_rows[1:3]:
        crew = row.find_all('td')
        crew_link = crew[0].find('a')

        crew_name, crew_url = crew_link.text, crew_link['href']
        crew_role = crew[-1].text
        crew_dict[crew_name] = {crew_role, crew_url}

    crew_id_dict[movie_id] = crew_dict 

crew_id_dict

mojo_df2 = pd.DataFrame(mojo_data_list)

mojo_df2.tail()

import json

mojo_df2.to_json('mojo_data2.json', orient='records', lines=True)

len(mojo_df)
