==> Project_2/05_webscraping_budget_numbers.ipynb <==
from bs4 import BeautifulSoup
import requests
import numpy as np
import pandas as pd

import os, re, sys, pickle, requests, resource, json
import time
from random import random
from IPython.core.display import clear_output

### Webscraping Movie Budgets and Domestic Grosses

For webscraping budget numbers, I used [The Numbers](https://www.the-numbers.com/).  The budget information can be accessed by individual movie page, or by a ranked list in the form of tables containing 100 titles at a time.


Although the list has a great many more titles, I chose initially to scrape the budget and gross information for the first 5000 titles in their ranking tables.  Since the ranking includes all genres, and not just science fiction, I assessed that this number would yield a sufficient number of records for a baseline linear regression.  I plan to do a larger scrape, including all budget and gross information available, at a future point.

The code below formats for the url for each of the 50 table pages, 100 titles each.  It then selects table information for title, release date, budget, domestic grosses, and worldwide grosses.

budget_data_list = []


budget_url = "https://www.the-numbers.com/movie/budgets/all/{}"

pages = [str(i) for i in range(1,5000,100)]

# preparing to monitor
start_time = time.time()
tries =0

for page in pages:

    response = requests.get(budget_url.format(page))
    source = response.text

    time.sleep(0.5+2*random())
    tries +=1
    elapsed_time = time.time() - start_time
    print('Request:{}; Frequency: {} requests/s'.format(tries, tries/elapsed_time))
    clear_output(wait = True)

    if tries > 2200:
        print("Number of requests exceeded")
        break

    budget_soup = BeautifulSoup(source, "lxml")

    budget_table = budget_soup.find('table')

    rows = [row for row in budget_table.find_all('tr')]

    for row in rows:
        items = row.find_all('td')
        if (len(items)<1) : 
            release_date = np.nan
            title = np.nan
            budget = np.nan
            dom_gross = np.nan
            ww_gross = np.nan
            
        else:
            release_date = items[1].text
            if not release_date: 
                next_element = item.findNext()

                if next_element:
                    release_date = next_element.text 
                else:
                    release_date = np.nan
            
            title = items[2].text
            if not title: 
                next_element = item.findNext()

                if next_element:
                    title = next_element.text 
                else:
                    title= np.nan 

            budget = items[3].text
            if not budget: 
                next_element = item.findNext()

                if next_element:
                    budget = next_element.text 
                else:
                    budget = np.nan

            dom_gross = items[4].text
            if not dom_gross: 
                next_element = item.findNext()

                if next_element:
                    dom_gross = next_element.text 
                else:
                    dom_gross = np.nan

            ww_gross = items[5].text
            if not ww_gross: 
                next_element = item.findNext()

                if next_element:
                    ww_gross = next_element.text 
                else:
                    ww_gross = np.nan

        budget_dict = dict(zip(headers, [title, release_date, budget, 
                                     dom_gross, ww_gross]))

        budget_data_list.append(budget_dict)

Checking that data was scraped appropriately for all records, converting the list of dictionaries to dataframe

len(budget_data_list)

len(budget_data_list[49])

budget_df = pd.DataFrame(budget_data_list)

budget_df.shape

budget_df = budget_df[budget_df['title'].notna()]

budget_df['title'].notna().sum()

Assessing to see that all intended categories of data were scraped

budget_df.sort_values(by="ww_gross", ascending=False)

#### Preliminary cleaning and formatting

The budget data will need to be merged with the IMDB and Rotten Tomatoes data.  Since I have already created a lowercase title string with underscores replacing spaces in order to format the Rotten Tomatoes url, I am formatting the budget data titles here to match this, to facilitate the merge.

budget_df['title_cc'] = budget_df['title'].str.lower()

budget_df['title_cc'] = budget_df['title_cc'].str.replace('&','and',regex=False)
budget_df['title_cc'] = budget_df['title_cc'].str.replace('\\','_',regex=True)
budget_df['title_cc'] = budget_df['title_cc'].str.replace('episode\s([ivx]*)\s-\s','',regex=True)
budget_df['title_cc'] = budget_df['title_cc'].str.replace("\s",'_',regex=True)
budget_df['title_cc'] = budget_df['title_cc'].str.replace("\W",'',regex=True)
budget_df['title_cc'] = budget_df['title_cc'].str.replace('_+','_',regex=True)

budget_df.head()

budget_df['budget'] = budget_df['budget'].str.strip('$')

budget_df['dom_gross'] = budget_df['dom_gross'].str.strip('$')

budget_df['ww_gross'] = budget_df['ww_gross'].str.strip('$')

budget_df.head()

budget_df.shape

budget_df['release_date2'] = budget_df['release_date'].replace('Unknown', np.nan)

budget_df['release_date2'] = budget_df['release_date2'].replace('Fall, 2021', np.nan)

#### Exporting and Saving Data

I exported the data in json format, for easy import into other notebooks for cleaning, merging, and analysis.

budget_df.to_json('budget_data.json', orient='records', lines=True)
